+++ platform_cli/groups/base.py
@@ -1,6 +1,7 @@
 import click
 
-class PlatformCliGroup():
+
+class PlatformCliGroup:
     def create(self, cli: click.Group):
         """Called by platform_cli, passing in the click group"""
-        pass 
+        pass