+++ platform_cli/groups/ros.py
@@ -7,7 +7,8 @@ import subprocess
 from platform_cli.groups.base import PlatformCliGroup
 from platform_cli.helpers import get_ros_env, echo, call
 
-def get_ros_poetry_packages(path:Path) -> List[Path]:
+
+def get_ros_poetry_packages(path: Path) -> List[Path]:
     package_xmls = glob(str(path / "**/package.xml"), recursive=True)
     package_xml_dirs = [Path(item).parent for item in package_xmls]
     pyproject_tomls = glob(str(path / "**/pyproject.toml"), recursive=True)
@@ -18,49 +19,54 @@ def get_ros_poetry_packages(path:Path) -> List[Path]:
         if dir in package_xml_dirs:
             ros_poetry_packages.append(dir)
 
-    echo(f"{len(ros_poetry_packages)} package(s) found", 'green')
+    echo(f"{len(ros_poetry_packages)} package(s) found", "green")
     return ros_poetry_packages
 
 
 class Ros(PlatformCliGroup):
-
     def create(self, cli: click.Group):
         @cli.group(help="CLI handlers associated with ROS packages")
         def ros():
             pass
-            
+
         @ros.command(name="build")
         @click.argument("args", nargs=-1)
-        def build(args: List[str]): # type: ignore
+        def build(args: List[str]):  # type: ignore
             """Runs colcon build on all ROS package"""
 
             env = get_ros_env()
             args_str = " ".join(args)
 
-            echo("Building packages...", 'green')
-            call(f"colcon build --merge-install --install-base /opt/greenroom/{env['PLATFORM_MODULE']} {args_str}")
+            echo("Building packages...", "green")
+            call(
+                f"colcon build --merge-install --install-base /opt/greenroom/{env['PLATFORM_MODULE']} {args_str}"
+            )
 
         @ros.command(name="test")
         @click.argument("args", nargs=-1)
-        def test(args: List[str]): # type: ignore
+        def test(args: List[str]):  # type: ignore
             """Runs colcon test on all ROS packages"""
 
             env = get_ros_env()
             args_str = " ".join(args)
 
-            echo("Testing packages...", 'green')
-            call(f"colcon test --merge-install --install-base /opt/greenroom/{env['PLATFORM_MODULE']} {args_str}")
+            echo("Testing packages...", "green")
+            call(
+                f"colcon test --merge-install --install-base /opt/greenroom/{env['PLATFORM_MODULE']} {args_str}"
+            )
             call("colcon test-result --all --verbose")
 
         @ros.command(name="install_poetry_deps")
-        @click.option('--base-path', type=str, help="The path to where the packages are installed")
-        def install_poetry_deps(base_path: Path): # type: ignore
+        @click.option("--base-path", type=str, help="The path to where the packages are installed")
+        def install_poetry_deps(base_path: Path):  # type: ignore
             """Installs the poetry deps for any python packages"""
 
             env = get_ros_env()
-            base_path = Path(base_path) if base_path else Path(f"/opt/greenroom/{env['PLATFORM_MODULE']}")
+            base_path = (
+                Path(base_path) if base_path else Path(f"/opt/greenroom/{env['PLATFORM_MODULE']}")
+            )
 
-            echo(f"Installing all poetry deps in {base_path} using pip", 'green')
+            echo(f"Installing all poetry deps in {base_path} using pip", "green")
             ros_poetry_packages = get_ros_poetry_packages(base_path)
             # Disable venv
             call("poetry config virtualenvs.create false")
@@ -69,7 +75,9 @@ class Ros(PlatformCliGroup):
                 echo(f"Installing {str(dir)}...", "blue")
                 # export dependencies to a requirements.txt file without hashes to decrease time to resolve dependencies.
                 # https://github.com/python-poetry/poetry-plugin-export/issues/78
-                call(f"poetry export -f requirements.txt --without-hashes --output requirements.txt", cwd=dir)
-                call(f"pip3 install -r requirements.txt", cwd=dir)
+                call(
+                    "poetry export -f requirements.txt --without-hashes --output requirements.txt",
+                    cwd=dir,
+                )
+                call("pip3 install -r requirements.txt", cwd=dir)
                 (dir / "requirements.txt").unlink()
-